+++ Sources/StreamChat/WebSocketClient/EventMiddlewares/EventDTOConverterMiddleware.swift
@@ -6,7 +6,7 @@ import Foundation
 
 /// The middleware responsible for converting event DTOs to events.
 struct EventDTOConverterMiddleware: EventMiddleware {
-    func handle(event: Event, session: DatabaseSession) -> Event? {
+    func handle(event: Event, session: DatabaseSession, notificationCenter: EventNotificationCenter) -> Event? {
         if let eventDTO = event as? EventDTO {
             return eventDTO.toDomainEvent(session: session)
         } else {