+++ Tests/StreamChatTests/WebSocketClient/EventMiddlewares/MemberEventMiddleware_Tests.swift
@@ -8,6 +8,7 @@ import XCTest
 
 final class MemberEventMiddleware_Tests: XCTestCase {
     var database: DatabaseContainer_Spy!
+    var center: EventNotificationCenter_Mock!
     var middleware: MemberEventMiddleware!
 
     // MARK: - Set up
@@ -16,12 +17,13 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         super.setUp()
 
         database = DatabaseContainer_Spy()
+        center = EventNotificationCenter_Mock(database: database)
         middleware = .init()
     }
 
     override func tearDown() {
-        AssertAsync.canBeReleased(&database)
         database = nil
+        AssertAsync.canBeReleased(&database)
         super.tearDown()
     }
 
@@ -31,7 +33,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let event = TestEvent()
 
         // Handle non-member event
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert event is forwarded as it is
         XCTAssertEqual(forwardedEvent as! TestEvent, event)
@@ -55,7 +57,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
 
         // Simulate and handle reaction event.
         let event = try MemberAddedEventDTO(from: eventPayload)
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert `MemberAddedEvent` is forwarded even though database error happened.
         XCTAssertTrue(forwardedEvent is MemberAddedEventDTO)
@@ -85,7 +87,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let channelListObserver = TestChannelListObserver(database: database)
 
         // Simulate `MemberAddedEvent` event.
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Load the channel.
         let channel = try XCTUnwrap(
@@ -139,7 +141,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         XCTAssertEqual(memberListQueryDTO?.members.map(\.user.id), [existingMember.user!.id])
 
         // Simulate `MemberAddedEventDTO` event.
-        _ = middleware.handle(event: event, session: database.viewContext)
+        _ = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert the new member is linked to the query
         XCTAssertEqual(memberListQueryDTO?.members.count, 2)
@@ -167,7 +169,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let event = try MemberAddedEventDTO(from: eventPayload)
 
         // WHEN
-        _ = middleware.handle(event: event, session: mockSession)
+        _ = middleware.handle(event: event, session: mockSession, notificationCenter: center)
 
         // THEN
         XCTAssertEqual(mockSession.markChannelAsReadParams?.cid, event.cid)
@@ -193,7 +195,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
 
         // Simulate and handle reaction event.
         let event = try MemberRemovedEventDTO(from: eventPayload)
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert `MemberRemovedEvent` is forwarded even though database error happened.
         XCTAssertTrue(forwardedEvent is MemberRemovedEventDTO)
@@ -252,7 +254,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let event = try MemberRemovedEventDTO(from: eventPayload)
 
         // Simulate `MemberRemovedEvent` event.
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Load the channel again
         channel = try XCTUnwrap(
@@ -299,7 +301,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
             createdAt: .unique
         )
         let event = try MemberRemovedEventDTO(from: eventPayload)
-        _ = middleware.handle(event: event, session: mockSession)
+        _ = middleware.handle(event: event, session: mockSession, notificationCenter: center)
 
         // THEN
         XCTAssertEqual(mockSession.markChannelAsUnreadParams?.cid, event.cid)
@@ -324,7 +326,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
 
         // Simulate and handle reaction event.
         let event = try MemberUpdatedEventDTO(from: eventPayload)
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert `MemberUpdatedEvent` is forwarded even though database error happened.
         XCTAssertTrue(forwardedEvent is MemberUpdatedEventDTO)
@@ -361,7 +363,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let event = try MemberUpdatedEventDTO(from: eventPayload)
 
         // Simulate `MemberUpdatedEvent` event.
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Load the channel again
         channel = try XCTUnwrap(
@@ -414,7 +416,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let channelListObserver = TestChannelListObserver(database: database)
 
         // Simulate `NotificationAddedToChannelEvent` event.
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert membership is not nil
         XCTAssertNotNil(channel)
@@ -470,7 +472,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         XCTAssertEqual(memberListQueryDTO?.members.map(\.user.id), [existingMember.user!.id])
 
         // Simulate `NotificationAddedToChannelEvent` event.
-        _ = middleware.handle(event: event, session: database.viewContext)
+        _ = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert the new member is linked to the query
         XCTAssertEqual(memberListQueryDTO?.members.count, 2)
@@ -510,7 +512,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let event = try NotificationRemovedFromChannelEventDTO(from: eventPayload)
 
         // Simulate `NotificationRemovedFromChannelEvent` event.
-        _ = middleware.handle(event: event, session: database.viewContext)
+        _ = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert membership is nil
         XCTAssertNotNil(channel)
@@ -555,7 +557,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let channelListObserver = TestChannelListObserver(database: database)
 
         // Simulate `NotificationAddedToChannelEvent` event.
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert membership is not nil
         XCTAssertNotNil(channel)
@@ -606,7 +608,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         XCTAssertEqual(memberListQueryDTO?.members.map(\.user.id), [existingMember.user!.id])
 
         // Simulate `NotificationInvitedEventDTO` event.
-        _ = middleware.handle(event: event, session: database.viewContext)
+        _ = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert the new member is linked to the query
         XCTAssertEqual(memberListQueryDTO?.members.count, 2)
@@ -648,7 +650,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let channelListObserver = TestChannelListObserver(database: database)
 
         // Simulate `NotificationAddedToChannelEvent` event.
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert membership is not nil
         XCTAssertNotNil(channel)
@@ -698,7 +700,7 @@ final class MemberEventMiddleware_Tests: XCTestCase {
         let channelListObserver = TestChannelListObserver(database: database)
 
         // Simulate `NotificationAddedToChannelEvent` event.
-        let forwardedEvent = middleware.handle(event: event, session: database.viewContext)
+        let forwardedEvent = middleware.handle(event: event, session: database.viewContext, notificationCenter: center)
 
         // Assert membership is not nil
         XCTAssertNotNil(channel)